# api/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: 'WornOut'
  description: 'Catalog API that links items and users together'
host: virtserver.swaggerhub.com
basePath: /aeckard87/wornOut/1.0
...
schemes:
  - http
paths:
...
  '/subcategories':
    get:
      tags:
        - Subcategories
      summary: Get all SubCategories
      operationId: getSubCategories
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/SubCategory'
        '404':
          description: No items found
    delete:
      tags:
        - Subcategories
      summary: Delete all SubCategories
      description: removes all subcategories.
      operationId: deleteSubCategories
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
  '/categories/{id}/subcategories':
    get:
      tags:
        - Subcategories
      summary: Get all SubCategories by Category
      operationId: getSubCategoriesByCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Category ID Associated to Subcategory(s)
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/SubCategory'
        '404':
          description: No items found
    post:
      tags:
        - Subcategories
      summary: Create a SubCategory by Category
      operationId: createSubCategoryByCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Category ID Associated to Subcategory(s)
        - in: body
          name: body
          description: Category that needs to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/SubCategory'
      responses:
        '201':
          description: SubCategory Created
          schema:
            $ref: '#/definitions/SubCategory'
        '400':
          description: Invalid input
    delete:
      tags:
        - Subcategories
      summary: Delete all SubCategories By Category
      description: removes all subcategories from a given category.
      operationId: deleteSubCategoriesByCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Category ID Associated to Subcategory(s)
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
  '/subcategories/{id}':
    get:
      tags:
        - Subcategories
      summary: Get a specific SubCategory
      operationId: getSubCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
      responses:
        '200':
          description: Status Ok
          schema:
            $ref: '#/definitions/SubCategory'
        '400':
          description: Invalid input
        '404':
          description: No items found
    put:
      tags:
        - Subcategories
      summary: Update a specific SubCategory
      operationId: updateSubCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
        - in: body
          name: body
          description: a subcategory that will be updated
          required: true
          schema:
            $ref: '#/definitions/SubCategory'
      responses:
        '201':
          description: SubCategory Updated
          schema:
            $ref: '#/definitions/SubCategory'
        '400':
          description: Invalid input
        '404':
          description: No items found
    delete:
      tags:
        - Subcategories
      summary: Delete a specific SubCategory
      description: removes a specfic subcategory.
      operationId: deleteSubCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
      responses:
        '200':
          description: Status Ok
        '400':
          description: Invalid input
        '404':
          description: No items found
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      operationId: getUsers
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '404':
          description: No items found
    post:
      tags:
        - Users
      summary: Create a User
      operationId: createUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: User that needs to be created
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: User Created
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid input
    delete:
      tags:
        - Users
      summary: Delete all Users
      operationId: deleteUsers
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
  '/users/{id}':
    get:
      tags:
        - Users
      summary: Get a specific User
      operationId: getUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
      responses:
        '200':
          description: Status Ok
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid input
        '404':
          description: No items found
    put:
      tags:
        - Users
      summary: Update a specific User
      operationId: updateUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
        - in: body
          name: body
          description: a user that needs to be updated
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: User Created
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid input
        '404':
          description: No items found
    delete:
      tags:
        - Users
      summary: Delete User
      operationId: deleteUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
      responses:
        '200':
          description: Status Ok
        '400':
          description: Invalid input
        '404':
          description: No items found
...
definitions:
...
  SubCategory:
    type: object
    properties:
      id:
        type: integer
        example: 290
      subcategory:
        type: string
        example: Short Sleeve
      categoryId:
        type: integer
        example: 10
...
  User:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
...